move the refactored JitImplCreatorBase class to a standalone file, NFC
