+++ deegen/CMakeLists.txt
@@ -55,6 +55,7 @@ add_library(deegen_utils
   deegen_simple_operand_expression.cpp
   deegen_analyze_read_locations.cpp
   deegen_jit_slow_path_data.cpp
+  deegen_jit_impl_creator_base.cpp
 ) 
 
 # Build the main Futamura projection stage binary
